alinux: locking/qspinlock/x86: Fix performance regression caused by virt_spin_lock
fix #31434803

commit 9043442b43b1 ("locking/paravirt: Use new static key for controlling call of virt_spin_lock()")
was intended to address the performance issue on baremetal machine
( or dedicated VM ). Because in these two situations, qspinlock is
preferred rather than trivial cmpxchg. The intention is good but
didn't apply to our environment.

9043442b431 uses KVM_HINTS_REALTIME to know if it's baremetal or dedicated.
But we uses KVM_FEATURE_PV_UNHALT to do the same thing.

So virt_spin_lock always gets to run, because it always treat our
environment as non-baremetal, non-dedicated. Huge performance loss
is resulted from this path.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>